requirements.txt
# YOLOv5 requirements for Jetson Nano
# Usage: pip install -r requirements.txt

# Base ------------------------------------------------------------------------
numpy>=1.19.0
opencv-python>=4.1.1
pillow>=7.1.2
PyYAML>=5.3.1
requests>=2.23.0
scipy>=1.4.1
tqdm>=4.64.0

# PyTorch should be installed separately for Jetson Nano
# Visit: https://forums.developer.nvidia.com/t/pytorch-for-jetson-version-1-10-now-available/72048

# Plotting (optional) --------------------------------------------------------
pandas>=1.1.4
seaborn>=0.11.0
matplotlib>=3.3.0

# Extras (optional) ----------------------------------------------------------
psutil  # system resources
thop>=0.1.1  # FLOPs computation

# Arduino communication ------------------------------------------------------
pyserial>=3.4  # 아두이노 통신
